+++ console/model/exceptions/InvalidGmailException.java
@@ -1,4 +1,4 @@
-package model.exceptions;
+package console.model.exceptions;
 
 /**
  * InvalidEmailException is an Exception class which is thrown when given email is not valid
